requirements.txt
altair
pandas
streamlit
python-dotenv
pymongo
